fix: SentenceGenerationのリネームと仕様変更
