--- src/components/general/SentenceGeneration.tsx
@@ -1,27 +1,0 @@
-import { Sentence } from '../../../types/Sentence';
-
-interface Props {
-  sentence: Sentence;
-  leading: string;
-  gap: string;
-  textSize: string;
-}
-
-export const SentenceGeneration: ({ gap, leading, sentence, textSize }: Props) => JSX.Element = ({
-  gap,
-  leading,
-  sentence,
-  textSize,
-}) => {
-  return (
-    <ul className={`grid font-main ${gap} ${textSize} ${leading} text-light`}>
-      {Object.values(sentence).map((paragraphs) => (
-        <li key={paragraphs.id}>
-          {Object.values(paragraphs.lines).map((texts) => (
-            <p key={texts}>{texts}</p>
-          ))}
-        </li>
-      ))}
-    </ul>
-  );
-};
